Simplified equality check
